[5]add combination of server side and client query
