+++ frontend/src/app/weather/page.tsx
@@ -1,18 +1,22 @@
-"use client";
-import { useQuery } from 'react-query'
+import Forcast from "./weatherforcast"
 
-function getForcast(){
+function getForcastDate(){
     return fetch('https://www.jma.go.jp/bosai/forecast/data/overview_forecast/130000.json')
-      .then(res => res.json())
+      .then(res => {
+        const json = res.json()
+        return json
+      }
+      )
 }
-
-export default function Weather() {
-  const { isLoading, error, data } = useQuery('forcast', getForcast)
+export default async function WeatherHome() {
+    const forcast_src = await getForcastDate()
     return (
       <div>
         <h1>Weather</h1>
-        <div>Forcast(Tokyo)</div>
-        <div>{isLoading ? "loading..." : data.text}</div>
+        <div>Forcast Of {forcast_src.reportDatetime} </div>
+        <div>*get in server side</div>
+        <Forcast></Forcast>
+        <div>*get in client</div>
       </div>
     )
   }
